requirements.txt
python-dotenv
openai
reportlab
beautifulsoup4
